requirements.txt
Flask
rembg
Pillow
onnxruntime
flask-cors
tensorflow